Moved events
https://coveord.atlassian.net/browse/JSUI-2693
